\name{alternativeSplicingEvents.hg19}
\alias{alternativeSplicingEvents.hg19}
\docType{package}
\title{
Alternative splicing event annotation for Human (hg19)
}
\description{
This package contains alternative splicing events compiled from multiple
programs. For more information, read the man pages of the following objects:
\itemize{
    \item{
        \code{\link{alternativeSplicingEvents.hg19.rda}} for Human (hg19
        assembly)
    }
}}
